+++ tools/generate_info.sh
@@ -1,8 +1,17 @@
 
+UNAMEOUT="$(uname -s)"
+case "${UNAMEOUT}" in
+    Linux*)    
+        date_info=$(date +"%Y-%m-%dT%H:%M:%S.%3NZ")
+        ;;
+    Darwin*)    
+        date_info=$(gdate +"%Y-%m-%dT%H:%M:%S.%3NZ")
+        ;;
+esac
+
 git_info=$(git describe --always --long --dirty --abbrev=10 --tags)
 git_short_hash=$(git rev-parse --short HEAD)
 git_hash=$(git rev-parse HEAD)
-date_info=$(date +"%Y-%m-%dT%H:%M:%S.%3NZ")
 
 echo "#ifndef __BUILD_INFO_H__"
 echo "#define __BUILD_INFO_H__"