+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 全屏模式下增加ESC退出全屏功能
 - 优化解决软件启动后固定占用CPU负载问题
 - 修复macos下全屏导致程序崩溃问题
+- 修复macos native UI模式下标题按钮没有切换为macos风格问题
 
 ## [[V0.4.1](https://github.com/QQxiaoming/quardCRT/releases/tag/V0.4.1)] - 2024-01-13